5.11

+++ src/main/java/synchronizer/verticles/storage/DeleteFileVerticle.java
@@ -1,10 +1,8 @@
 package synchronizer.verticles.storage;
 
 import io.vertx.core.AbstractVerticle;
-import io.vertx.core.Future;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import synchronizer.exceptions.VerticleException;
 
 import java.io.File;
 
@@ -29,12 +27,17 @@ public class DeleteFileVerticle extends AbstractVerticle {
 
     @Override
     public void start(){
+        if (fileToDelete == null || fileToDelete.isEmpty()){
+            logger.warn(String.format("received invalid file to delete %s", fileToDelete));
+            return;
+        }
         // Check existence and delete
         vertx.fileSystem().exists(fileToDelete, result -> {
             if (result.succeeded() && result.result()) {
                 vertx.fileSystem().deleteBlocking(fileToDelete);
+                logger.info(String.format("deleted file %s", fileToDelete));
             } else {
-                logger.info(String.format("Failed to remove file %s", fileToDelete));
+                // ignore deletion because file already don't exist
             }
         });
     }
